pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools.backends"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "confab_test"
version = "0.1.0"
description = "Automated confabulation test suite for local LLMs via Ollama"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "rich>=13",
    "click>=8",
    "pyyaml>=6",
]

[project.scripts]
confab-test = "confab_test.cli:main"

[tool.setuptools.packages.find]
where = ["."]
